fix(stream): fix error.

+++ include/libs/stream/tstream.h
@@ -656,7 +656,7 @@ typedef struct STaskStatusEntry {
   STaskId id;
   int32_t status;
   int32_t statusLastDuration; // to record the last duration of current status
-  int32_t stage;
+  int64_t stage;
   int32_t nodeId;
   int64_t verStart;         // start version in WAL, only valid for source task
   int64_t verEnd;           // end version in WAL, only valid for source task
